rag_project/requirements.txt
# RAG System — Complete Requirements
# Run: pip install -r requirements.txt

# ── Core ML / NLP ──────────────────────────────────────────────────────────────
torch==2.2.0
transformers==4.40.0
sentence-transformers==2.7.0
datasets==2.19.0
accelerate==0.29.3

# ── PEFT / Fine-Tuning ─────────────────────────────────────────────────────────
peft==0.10.0
trl==0.8.6

# ── Vector Databases ───────────────────────────────────────────────────────────
faiss-cpu==1.8.0
chromadb==0.5.0

# ── Evaluation ─────────────────────────────────────────────────────────────────
nltk==3.8.1
rouge-score==0.1.2

# ── PDF Support ────────────────────────────────────────────────────────────────
PyMuPDF==1.24.0

# ── UI & Utilities ─────────────────────────────────────────────────────────────
streamlit==1.33.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
